call: add ModemCall:start-time property
The call property mapper which maps properties of o.o.VoiceCall to
ModemCall properties contains a "start-time" property which doesn't
actually exist on ModemCall. Let's add it. It's currently not the most
useful property ever though.

The other fix would be to remove the map from StartTime to start-time
from the property mapper.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>