ring-conference: reimplement conf channels with TpBaseChannel
A new implementation of conference channels using the updated modem
interface (ModemCallConference not used anymore).

Instead of inheriting RingMediaChannel, conference channels are now
implemented directly using TpBaseChannel. This design better matches
the underlying oFono APIs, where conference call is not a special
case and/or extension of call, but a totally separate interface.