ring-call-channel.c: fix monitoring of multiparty status
The modem no longer sends a 'multiparty' signal, but instead
'notify::multiparty' should be used.

The old signal was related to remote party merging the call into
a conference. This information is no longer available through
the modem interface. The new 'notify::multiparty' interface relates
to membership in locally hosted conferences. With the new modem
interface, client needs to monitor this signal to monitor for
conference splits.