[telepathy-ring] Always include change-reason in MembersChangedDetailed
A Channel_Group_Change_Reason of 0 ("None") is valid, and according to
spec should be used to indicate e.g. normal termination by the remote
user.

TelepathyQt reacts differently to a signal with no change-reason field:
it assumes that the signal was not detailed enough to indicate the
cause. This eventually results in a channel close with a vague error
that is misinterpreted as a local hangup instead of a remote one, causing
calls that should be considered missed to not be flagged as such.